+++ src/system/device.js
@@ -4,6 +4,7 @@ import save from "./save.js";
 import { prefixed } from "./../utils/agent.js";
 import state from "./../state/state.js";
 import * as event from "./event.js";
+import { DOMContentLoaded } from "./dom.js";
 
 // private properties
 let accelInitialized = false;
@@ -21,45 +22,6 @@ function _disableSwipeFn(e) {
         globalThis.scroll(0, 0);
     }
     return false;
-};
-
-// DOM loading stuff
-let readyBound = false, isReady = false, readyList = [];
-
-/**
- * // called to check if the device is ready
- * @ignore
- */
-function _domReady() {
-    // Make sure that the DOM is not already loaded
-    if (!isReady) {
-        // be sure document.body is there
-        if (typeof globalThis.document !== "undefined" && !globalThis.document.body) {
-            return setTimeout(_domReady, 13);
-        }
-
-        // clean up loading event
-        if (typeof globalThis.document !== "undefined" && typeof globalThis.document.removeEventListener === "function") {
-            globalThis.document.removeEventListener(
-                "DOMContentLoaded",
-                this._domReady,
-                false
-            );
-        }
-
-        if (typeof globalThis.removeEventListener === "function") {
-            // remove the event on globalThis.onload (always added in `onReady`)
-            globalThis.removeEventListener("load", _domReady, false);
-        }
-
-        // execute all callbacks
-        while (readyList.length) {
-            readyList.shift().call(globalThis, []);
-        }
-
-        // Remember that the DOM is ready
-        isReady = true;
-    }
 };
 
 // a cache DOMRect object
@@ -606,33 +568,7 @@ let device = {
     * });
     */
     onReady(fn) {
-        // If the DOM is already ready
-        if (isReady) {
-            // Execute the function immediately
-            fn.call(globalThis, []);
-        }
-        else {
-            // Add the function to the wait list
-            readyList.push(fn);
-
-            // attach listeners if not yet done
-            if (!readyBound) {
-                // directly call domReady if document is already "ready"
-                if (device.nodeJS === true || (typeof globalThis.document !== "undefined" && globalThis.document.readyState === "complete")) {
-                    // defer the fn call to ensure our script is fully loaded
-                    globalThis.setTimeout(_domReady, 0);
-                }
-                else {
-                    if (typeof globalThis.document !== "undefined" && typeof globalThis.document.addEventListener === "function") {
-                        // Use the handy event callback
-                        globalThis.document.addEventListener("DOMContentLoaded", _domReady, false);
-                    }
-                    // A fallback to globalThis.onload, that will always work
-                    globalThis.addEventListener("load", _domReady, false);
-                }
-                readyBound = true;
-            }
-        }
+        DOMContentLoaded(fn);
     },
 
     /**